requirements.txt
assertpy==1.1
attrs==22.2.0
behave==1.2.6
behave-restful==0.4.4
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
Jinja2==3.1.2
jproperties==2.1.1
jsonpath==0.82
jsonref==1.1.0
jsonschema==4.17.3
MarkupSafe==2.1.2
parse==1.19.0
parse-type==0.6.0
pyrsistent==0.19.3
PyYAML==6.0
requests==2.28.2
six==1.16.0
urllib3==1.26.14
requests_to_curl